raw_models/khelbot-para-athlete/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv